Create Simple parser for fish data.
Data was extracted (simply copy-paste) from the fandom site.

The short script just formats it to my 1st-phase datasructure.

It's really ugly right now, I know that, I will likely modify this soon.

Also, one thing to consider would be to move this data to an
S3 bucket. bundling the data to the app wouldn't be ideal as any
updates to the core data would require an entire update to the app.

I think it's okay for now, but again, consider doing the work to
move it future Noel