Keep track of active header cell.
This is intended to make sure that we can select
a header, and press it to influence the sort state
of the table.

I ended up creating a separate component for the
tableHeader since I'll have to reuse this in
the other tables.

so far it does nothing other than colour the cell