Implement sorting functionality for table
I also moved some code around to make a utils area.
I didn't think it made sense to keep it in the root
level so I started a src folder and created a utils folder.

Also, I know that the sorting code is super rough,
This is the first version so I'll just keep it as is for now.