Fix bottom safe area insets
This is hacky, this shouldn't be the way you do this,
realistically, you should only need to use the bottom inset
in the bottomPadding, but for some reason, in our
set up, it requires both the top and the bottom.

Look into why this is the case. Fix it Noel!!