Update hemisphere context values to be ready for display
I know this may not be the ideal way to do this, since I am
now relying on code to dictate what to display, but I am
just experimenting what to do since I don't want to have to
do the same logic all over the place when trying to determine
what to display when the hemisphere is north or south.

I actually think this may be more trouble than it's worth especially
when I consider adding proper localization.