Create a right header area context
This was necessary due to the nature of how I am
currently using the navigation stack.

Due to the fact that I'm embedding the drawer anvigator
into the stack navigator, the flow of options from pages
don't propagate properly all the way up to the stack navigator.

From my understanding based on my readings from the
react/navigation documentation, how useNavigation() works
in an embedded scenario, is that the useNavigation hook
will return you the closest parent to the page, so in my case,
the drawer's navigation object. If i immediately call setOptions
on that, the drawerNavigator swallows up the options, and the
rightHeader buttons never propagate all the way up to the
stack navigator.

To mitigate this, I created my own hooks to set and read out
the header buttons (the pages set it, and the stack navigator
reads it into its options.)

I know that this is super convoluted, but this is the approach that
ended up working for me.

If I find a better solution, I will consider scraping this change.

This also got me thinking that there may be a better framework
for me to use that solves this problem for me already....